+++ src/FilamentOnboardingServiceProvider.php
@@ -32,6 +32,7 @@ class FilamentOnboardingServiceProvider extends PackageServiceProvider
                 'add_media_to_onboarding_steps',
                 'add_visibility_to_onboarding',
                 'harden_onboarding_progress_scope',
+                'create_onboarding_preferences',
             ])
             ->hasCommand(ResetOnboardingCommand::class);
     }